Added enums and started adding to interface
Created enums based on Opcodes from ASM file. Used to make it simplier
on us when retrieving access levels. If discovered that they are not
needed, they can be removed. Thought it would be a nice touch to make it
easier to read the code and understand the access levels. Also started
adding methods to interface, only field so far. No errors so far. Plenty
of TODOs.