Cleaned up output file to output nicely
Removed some code that made it cleaner in the png, but apparently it is
not needed, even though it looks nice. Aaron will probably never read
this, so he won't that I had some sort of anger towards that. Also,
output to graph.gv file is human readable for now and is helpful for
debugging. Have to fix a couple of things.