Added panels after parser is run
When it is finished analyzing, it now shows a new panel with a checkbox
panel on the left and image panel on the right. However it does not
display anything right now. I had a lot of trouble making a proxy that
waits for grapviz to give us a png.